jb-bot/pyproject.toml
[tool.poetry]
name = "jb-bot"
version = "0.1.0"
description = "A tool for querying your JBrowse configuration file."
authors = ["Caroline Bridge <cbridge@oicr.on.ca>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
chromadb = "^0.5.11"
flask = "^3.0.3"
flask-cors = "^5.0.0"
grpcio = "1.60.1"
ipython = "^8.27.0"
jsonpath-ng = "^1.6.1"
langchain = "^0.3.1"
langchain-chroma = "^0.1.4"
langchain-community = "^0.3.1"
langchain-google-genai = "^2.0.1"
langchain-openai = "^0.2.1"
langchain-text-splitters = "^0.3.0"
openai = "^1.51.0"
python = "^3.12"
python-dotenv = "^1.0.1"
rich = "^13.9.2"
sqlite-utils = "^3.37"
langgraph = "^0.2.39"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
